requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
ca-certificates=2020.12.5=ha878542_0
certifi=2020.12.5=py37h89c1867_1
cffi=1.14.4=py37h261ae71_0
cudatoolkit=10.0.130=0
cycler=0.10.0=py_2
dbus=1.13.6=he372182_0
expat=2.2.9=he1b5a44_2
fontconfig=2.13.1=he4413a7_1000
freetype=2.10.4=h5ab3b9f_0
glib=2.66.1=h92f7085_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=hf484d3e_1000
intel-openmp=2020.2=254
joblib=1.0.0=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=py37hc928c03_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libllvm10=10.0.1=he513fc3_3
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=h8b12597_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=2.32.1=h14c3975_1000
libxcb=1.13=h14c3975_1002
libxml2=2.9.10=hb55368b_3
llvmlite=0.34.0=py37h269e1b5_4
lz4-c=1.9.3=h2531618_0
matplotlib=3.3.3=py37h89c1867_0
matplotlib-base=3.3.3=py37h4f6019d_0
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.0.4=py37hd81dba3_0
ncurses=6.2=he6710b0_1
ninja=1.10.2=py37hff7bd54_0
numba=0.51.2=py37h9fdb41a_0
numpy=1.19.1=py37hbc911f0_0
numpy-base=1.19.1=py37hfa32c7d_0
olefile=0.46=py37_0
openssl=1.1.1i=h27cfd23_0
pcre=8.44=he1b5a44_0
pillow=8.1.0=py37he98fc37_0
pip=20.3.3=py37h06a4308_0
protobuf=3.13.0.1=py37he6710b0_1
pthread-stubs=0.4=h36c2ea0_1001
pycparser=2.20=py_2
pynndescent=0.5.1=pyhd3deb0d_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py37hcca6a23_4
python=3.7.9=h7579374_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.0.1=py3.7_cuda10.0.130_cudnn7.4.2_2
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
scikit-learn=0.23.2=py37h0573a6f_0
scipy=1.5.2=py37h0b6359f_0
setuptools=52.0.0=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py_0
sqlite=3.33.0=h62c20be_0
tbb=2020.2=hc9558a2_0
tensorboardx=2.1=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
torchvision=0.2.2=py_3
tornado=6.1=py37h4abf009_0
umap-learn=0.5.0=py37h89c1867_0
wheel=0.36.2=pyhd3eb1b0_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
